+++ axlearn/common/flash_attention/tpu_attention_benchmark.py
@@ -1,6 +1,17 @@
 # Copyright © 2023 Apple Inc.
 
 """Benchmark TPU FlashAttention kernels.
+For fast running kernel, python benchmark time may not be accurate.
+Therefore, we tyically would enable jax profiler checking trace.
+Example run:
+    python tpu_attention_benchmark.py
+            --config=134b \
+            --run_reference=False \
+            --enable_trace_profiling=True \
+            --batch_size=8 \
+            --seq_len=65536 \
+            --page_size=128
+    Traced log dir can be specified with --trace-dir.
 
 Sample outputs: (v5p)
 CMD: python \
@@ -30,6 +41,7 @@ import time
 from typing import Callable, Optional
 
 import jax
+from absl import app, flags
 
 from axlearn.common.attention_bias import causal_mask
 from axlearn.common.flash_attention.common import ReferenceMHA
@@ -41,6 +53,17 @@ from axlearn.common.flash_attention.utils import flash_attention_implementation
 from axlearn.common.kv_cache.base_kv_cache import BaseKVCache
 from axlearn.common.kv_cache.kv_cache import KVCache
 from axlearn.common.kv_cache.paged_kv_cache import PagedKVCache
+
+flags.DEFINE_boolean("enable_trace_profiling", False, "Whether to enable JAX trace profiling.")
+flags.DEFINE_string("trace_dir", "/tmp/axlearn_profiler", "Directory to save profiler traces.")
+flags.DEFINE_boolean("run_reference", True, "Whether to run the reference implementation.")
+flags.DEFINE_string("config", None, "Configuration to run benchmark.")
+flags.DEFINE_integer("batch_size", 2, "Batch size for running benchmark.")
+flags.DEFINE_integer("seq_len", 8192, "Sequence length to run benchmark.")
+flags.DEFINE_integer("page_size", None, "Page size for paged attention.")
+flags.DEFINE_integer("sliding_window_size", None, "Sliding window size for attention.")
+FLAGS = flags.FLAGS
+
 
 _BENCHMARK_CONFIGS = {
     "1.2b": dict(
@@ -69,10 +92,10 @@ _BENCHMARK_CONFIGS = {
         per_head_dim=128,
     ),
     # OOM in mha_reference.
-    # "539.5b": dict(
-    #     num_heads=140,
-    #     per_head_dim=128,
-    # ),
+    "539.5b": dict(
+        num_heads=140,
+        per_head_dim=128,
+    ),
 }
 
 
@@ -80,8 +103,12 @@ def _time_call(fn: Callable, *, num_iters: int = 5) -> float:
     """Times average execution time for fn call over num_iters after warmup."""
     fn().block_until_ready()
     tic = time.perf_counter()
+    if FLAGS.enable_trace_profiling:
+        jax.profiler.start_trace(FLAGS.trace_dir)
     for _ in range(num_iters):
         fn().block_until_ready()
+    if FLAGS.enable_trace_profiling:
+        jax.profiler.stop_trace()
     toc = time.perf_counter()
     return (toc - tic) / num_iters
 
@@ -103,7 +130,8 @@ def _benchmark(
     """Benchmarks TPU FlashAttention vs reference impl."""
     if not (kv_cache_type is None or kv_cache_type in (KVCache, PagedKVCache)):
         raise NotImplementedError(f"This benchmark doesn't support {kv_cache_type=} yet.")
-    if kv_cache_type == PagedKVCache:
+    if page_size is not None:
+        kv_cache_type = PagedKVCache
         assert page_size is not None
         q, k, v, page_tables, bias = generate_paged_attention_data(
             batch_size=batch_size,
@@ -154,39 +182,61 @@ def _benchmark(
     )
 
     input_batch = dict(query=q, key=k, value=v, bias=bias, page_tables=page_tables)
-    ref_fwd_time = _time_call(lambda: ref_mha_impl(input_batch))
+    if FLAGS.run_reference:
+        ref_mha_impl = (
+            ReferenceMHA.default_config()
+            .set(softmax_scale=softmax_scale, tpu_block_size=block_size)
+            .instantiate()
+        )
+        ref_fwd_time = _time_call(lambda: ref_mha_impl(input_batch))
+        print(f"ref_fwd: {ref_fwd_time:.4f}s")
+
     flash_fwd_time = _time_call(lambda: mha_impl(input_batch))
-    print(f"ref_fwd:{ref_fwd_time:.4f}s, flash_fwd:{flash_fwd_time:.4f}s")
+    print(f"flash_fwd:{flash_fwd_time:.4f}s")
 
     if kv_cache_type is None:
-
-        def grad_ref(float_inputs, aux_inputs):
-            full_batch = {**float_inputs, **aux_inputs}
-            return ref_mha_impl(full_batch).mean()
 
         def grad_test(float_inputs, aux_inputs):
             full_batch = {**float_inputs, **aux_inputs}
             return mha_impl(full_batch).mean()
 
+        if FLAGS.run_reference:
+
+            def grad_ref(float_inputs, aux_inputs):
+                full_batch = {**float_inputs, **aux_inputs}
+                return ref_mha_impl(full_batch).mean()
+
+            ref_grad_fn = jax.jit(jax.grad(grad_ref, argnums=0))
+            ref_bwd_time = _time_call(lambda: ref_grad_fn(float_inputs, aux_inputs)["query"])
+            print(f"ref_bwd:{ref_bwd_time:.4f}s")
+
         float_inputs = dict(query=q, key=k, value=v)
         aux_inputs = dict(bias=bias)
-        ref_grad_fn = jax.jit(jax.grad(grad_ref, argnums=0))
-        ref_bwd_time = _time_call(lambda: ref_grad_fn(float_inputs, aux_inputs)["query"])
         flash_grad_fn = jax.jit(jax.grad(grad_test, argnums=0))
         flash_bwd_time = _time_call(lambda: flash_grad_fn(float_inputs, aux_inputs)["query"])
-        print(f"ref_bwd:{ref_bwd_time:.4f}s, flash_bwd:{flash_bwd_time:.4f}s\n")
+        print(f"flash_bwd:{flash_bwd_time:.4f}s")
 
 
-if __name__ == "__main__":
+def main(_):
     assert jax.default_backend() == "tpu", "Benchmarking requires a TPU backend."
     device_kind = jax.devices()[0].device_kind
-    for name, cfg in _BENCHMARK_CONFIGS.items():
+    if FLAGS.config is not None:
+        config_list = [(FLAGS.config, _BENCHMARK_CONFIGS[FLAGS.config])]
+    else:
+        # Run Sweep without benchmarking 539.5b
+        config_list = [(k, v) for k, v in _BENCHMARK_CONFIGS.items() if k != "539.5b"]
+    for name, cfg in config_list:
         print(f"Benchmarking attention representative of {name} model layer on {device_kind}.")
         _benchmark(
-            batch_size=2,
-            seq_len=1024 * 8,
+            batch_size=FLAGS.batch_size,
+            seq_len=FLAGS.seq_len,
             block_size=4 * 128,
-            sliding_window_size=4096,
+            sliding_window_size=FLAGS.sliding_window_size,
+            page_size=FLAGS.page_size,
             kv_cache_type=KVCache,
             **cfg,
         )
+
+
+if __name__ == "__main__":
+    app.run(main)
