+++ axlearn/common/flash_attention/tpu_decoding.py
@@ -36,16 +36,10 @@ from axlearn.common.attention_bias import (
     BaseAttentionBias,
     MaskFn,
     MaskFnAttentionBias,
-    SlidingWindowAttentionBias,
     split,
 )
-from axlearn.common.flash_attention.common import (
-    BaseSingleStepDecoding,
-    build_mask,
-    build_sliding_window_mask,
-    get_tpu_dot_precision,
-    query_iterator_indices,
-)
+from axlearn.common.flash_attention.common import BaseSingleStepDecoding, get_tpu_dot_precision
+from axlearn.common.flash_attention.tpu_paged_attention_kernel import prepare_block_sparse_map
 from axlearn.common.kv_cache.base_kv_cache import BaseKVCache
 from axlearn.common.kv_cache.kv_cache import KVCache
 from axlearn.common.utils import Nested, Tensor
@@ -202,43 +196,12 @@ class TPUDecoding(BaseSingleStepDecoding):
         v = jnp.einsum("bsnh->bnhs", value)
         bs, kv_heads, head_dim, padded_kv_seq_len = k.shape
         kv_seq_len = jnp.broadcast_to(jnp.asarray(kv_seq_len), (bs,))
-        # Computes a full block map num_kv_blocks * num_kv_blocks.
-        # Use a padding to ensure padding blocks aren't counted towards `kv_block_offset_size`.
-        padding = -1
-        with jax.ensure_compile_time_eval():
-            if mask_fn is not None:
-                mask_args = dict(
-                    q_seq_len=padded_kv_seq_len,
-                    kv_seq_len=padded_kv_seq_len,
-                    block_q=block_size,
-                    block_k=block_size,
-                )
-                if isinstance(mask, SlidingWindowAttentionBias):
-                    bool_mask = build_sliding_window_mask(
-                        **mask_args, sliding_window_size=mask.sliding_window_size
-                    )
-                else:
-                    bool_mask = build_mask(mask_fn, **mask_args)
-                offset, _ = query_iterator_indices(bool_mask, padding=padding)
-            else:
-                padded_num_kv_blocks = pl.cdiv(padded_kv_seq_len, block_size)
-                offset = lax.broadcasted_iota(
-                    jnp.int32, (padded_num_kv_blocks, padded_num_kv_blocks), 1
-                )
-
-        # Dynamically slice the rows according to the query position (which is kv_seq_len - 1).
-        kv_block_offset = offset[(kv_seq_len - 1) // block_size]
-        # Count the number of blocks with position < kv_seq_len.
-        kv_block_offset_size = jnp.count_nonzero(
-            (kv_block_offset != padding) & (kv_block_offset * block_size < kv_seq_len[:, None]),
-            axis=1,
+        kv_block_offset, kv_block_offset_size = prepare_block_sparse_map(
+            mask,
+            lengths=kv_seq_len,
+            block_size=block_size,
+            seq_len=padded_kv_seq_len,
         )
-        # Replace padding with the last valid kv block's index. See
-        # https://docs.jax.dev/en/latest/pallas/tpu/sparse.html#sparse-access-patterns-on-dense-data
-        kv_block_offset = jnp.where(
-            kv_block_offset == padding, kv_block_offset.max(axis=1, keepdims=True), kv_block_offset
-        )
-
         q = q.reshape(bs, kv_heads, -1, head_dim)
         q_seq_head = q.shape[-2]  # = q_seq_len * num_q_heads_per_kv_head
         assert q_seq_head <= 512
